+++ src/controllers/mod.rs
@@ -535,6 +535,50 @@ impl InfoAggiuntiveController {
         state.infoaggiuntive_model.set_valid(false);
     }
 
+    pub fn check_larghezza_stazione_string(&self, larghezza: &str) -> Result<f32,String> {
+        let s = larghezza.replace(',', "."); // Replace comma with dot
+        match s.parse::<f32>() {
+            Ok(value) => {
+                return Ok(value);
+            }
+            Err(e) => {
+                let mut err_msg = format!("Errore nella conversione larghezza stazione: {}", e);
+                if err_msg.contains("invalid float literal") {
+                    err_msg = err_msg.replace("invalid float literal", "tipo non valido: atteso razionale");
+                }
+                self.add_console_message(format!("InfoAggiuntiveController:  {err_msg}"));
+                let mut state = GLOBAL_STATE.lock().unwrap();
+                state.data_model.set_anagrafica_niseci(None);
+                state.infoaggiuntive_model.set_done_editing(false);
+                state.infoaggiuntive_model.set_valid(false);
+                state.infoaggiuntive_model.set_errors_occurred(true);
+                return Err(err_msg);
+            }
+        }
+    }
+
+    pub fn check_lunghezza_stazione_string(&self, lunghezza: &str) -> Result<f32,String> {
+        let s = lunghezza.replace(',', "."); // Replace comma with dot
+        match s.parse::<f32>() {
+            Ok(value) => {
+                return Ok(value);
+            }
+            Err(e) => {
+                let mut err_msg = format!("Errore nella conversione lunghezza stazione: {}", e);
+                if err_msg.contains("invalid float literal") {
+                    err_msg = err_msg.replace("invalid float literal", "tipo non valido: atteso razionale");
+                }
+                self.add_console_message(format!("InfoAggiuntiveController:  {err_msg}"));
+                let mut state = GLOBAL_STATE.lock().unwrap();
+                state.data_model.set_anagrafica_niseci(None);
+                state.infoaggiuntive_model.set_done_editing(false);
+                state.infoaggiuntive_model.set_valid(false);
+                state.infoaggiuntive_model.set_errors_occurred(true);
+                return Err(err_msg);
+            }
+        }
+    }
+
     pub fn valida_anagrafica_niseci(&self) {
 
         //We grab the state in a scope to ensure we don't get lock problems