+++ CHANGELOG.md
@@ -15,6 +15,9 @@
   - Refactor field `numPassaggio` to be `u32`
 - Refactor `RecordCsvAnagraficaNISECI`:
   - Refactor fields `larghezza_stazione`, `lunghezza_stazione` to be `f32`
+  - Refactor `SelezioneInfoAggiuntiveView` to handle the check
+    - ATM it's done by `InfoAggiuntiveController::check_larghezza_stazione_str()`
+    - Also updates the model and sets the error flag, going to `ConsoleView`
 - Flag `--info` now also prints expected types
 - Fixed wrong message in error reports for `check_riferimento_niseci_reader`